backend/requirements.txt
fastapi
motor
uvicorn
pydantic
python-jose[cryptography]
python-dotenv
passlib
fpdf
reportlab